+++ trunk/ContraGame/ContraGame/CBBullet.h
@@ -14,7 +14,7 @@ public:
 	~CBBullet();
 
 	void LoadResources();
-	void Update(int delta_time);
+	void Update(float delta_time);
 	void Draw();
 
 	void Moving(float v_max);
